+++ sdk/spring/spring-cloud-azure-autoconfigure/src/main/java/com/azure/spring/cloud/autoconfigure/aad/AadResourceServerWebSecurityConfigurerAdapter.java
@@ -10,9 +10,14 @@ import org.springframework.core.convert.converter.Converter;
 import org.springframework.security.authentication.AbstractAuthenticationToken;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.configurers.oauth2.server.resource.OAuth2ResourceServerConfigurer;
+import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationConverter;
+import org.springframework.util.Assert;
 import org.springframework.util.StringUtils;
+
+import java.util.Collection;
 
 /**
  * Abstract configuration class, used to make JwtConfigurer and AADJwtBearerTokenAuthenticationConverter take effect.
@@ -22,10 +27,32 @@ import org.springframework.util.StringUtils;
 public abstract class AadResourceServerWebSecurityConfigurerAdapter extends WebSecurityConfigurerAdapter {
 
     @Autowired
-    AadResourceServerProperties properties;
+    private AadResourceServerProperties properties;
+
+    private Converter<Jwt, Collection<GrantedAuthority>> jwtGrantedAuthoritiesConverter;
 
     /**
-     * configure
+     * Creates a new instance with the default configuration.
+     */
+    public AadResourceServerWebSecurityConfigurerAdapter() {
+    }
+
+    /**
+     * Sets the Azure AD properties and custom granted authority converter to creates a new instance,
+     * the custom granted authority converter can be null.
+     *
+     * @param properties the Azure AD properties for Resource Server
+     * @param jwtGrantedAuthoritiesConverter the custom converter for JWT granted authority
+     */
+    public AadResourceServerWebSecurityConfigurerAdapter(AadResourceServerProperties properties,
+                                                         Converter<Jwt, Collection<GrantedAuthority>> jwtGrantedAuthoritiesConverter) {
+        Assert.notNull(jwtGrantedAuthoritiesConverter, "jwtGrantedAuthoritiesConverter cannot be null");
+        this.properties = properties;
+        this.jwtGrantedAuthoritiesConverter = jwtGrantedAuthoritiesConverter;
+    }
+
+    /**
+     * Apply the {@link OAuth2ResourceServerConfigurer}  for Azure AD OAuth2 Resource Server scenario.
      *
      * @param http the {@link HttpSecurity} to use
      * @throws Exception Configuration failed
@@ -45,8 +72,22 @@ public abstract class AadResourceServerWebSecurityConfigurerAdapter extends WebS
         if (StringUtils.hasText(properties.getPrincipalClaimName())) {
             converter.setPrincipalClaimName(properties.getPrincipalClaimName());
         }
-        converter.setJwtGrantedAuthoritiesConverter(
-            new AadJwtGrantedAuthoritiesConverter(properties.getClaimToAuthorityPrefixMap()));
+
+        this.jwtGrantedAuthoritiesConverter = jwtGrantedAuthoritiesConverter();
+        if (this.jwtGrantedAuthoritiesConverter != null) {
+            converter.setJwtGrantedAuthoritiesConverter(jwtGrantedAuthoritiesConverter);
+        } else {
+            converter.setJwtGrantedAuthoritiesConverter(
+                new AadJwtGrantedAuthoritiesConverter(properties.getClaimToAuthorityPrefixMap()));
+        }
         return converter;
+    }
+
+    /**
+     * Customize the Jwt granted authority converter, and return the {@link AadResourceServerWebSecurityConfigurerAdapter#jwtGrantedAuthoritiesConverter} by default.
+     * @return the Jwt granted authority converter.
+     */
+    protected Converter<Jwt, Collection<GrantedAuthority>> jwtGrantedAuthoritiesConverter() {
+        return this.jwtGrantedAuthoritiesConverter;
     }
 }