+++ sdk/spring/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 4.5.0 (2022-12-05)
+## 4.5.0 (2022-12-06)
 Upgrade Spring Boot dependencies version to 2.7.6 and Spring Cloud dependencies version to 2021.0.5
 Upgrade Spring Boot dependencies version to 2.7.4 and Spring Cloud dependencies version to 2021.0.4
 
@@ -8,8 +8,8 @@ Upgrade Spring Boot dependencies version to 2.7.4 and Spring Cloud dependencies 
 This section includes changes in `spring-cloud-azure-starter` module.
 
 #### Features Added
-- GA the `spring-cloud-azure-starter-jdbc-mysql`. This starter supports Azure hosted MySQL service authenticating with Azure AD[#30024](https://github.com/Azure/azure-sdk-for-java/pull/30024).
-- GA the `spring-cloud-azure-starter-jdbc-postgresql`. This starter supports Azure hosted PostgreSQL service authenticating with Azure AD[#30024](https://github.com/Azure/azure-sdk-for-java/pull/30024).
+- GA the `spring-cloud-azure-starter-jdbc-mysql`. This starter supports Azure hosted MySQL service authenticating with Azure AD.
+- GA the `spring-cloud-azure-starter-jdbc-postgresql`. This starter supports Azure hosted PostgreSQL service authenticating with Azure AD.
 
 ### Spring Cloud Azure Autoconfigure
 This section includes changes in `spring-cloud-azure-autoconfigure` module.